Restructuring and cleaning code

+++ sipsl-code/hdr/UTIL.h
@@ -81,7 +81,7 @@ typedef struct {
 				MESSAGE* __mess = new MESSAGE(__echob, __sode, inTime, __sock, __echoAddr);\
 				DEBOUT("NEW MESSAGE"," " + __mess->getTotLines());\
 				long long int num = ((long long int) inTime.tv.tv_sec)*1000000+(long long int)inTime.tv.tv_usec;\
-				sprintf(bu, "%x#%llu",__mess,num);\
+				sprintf(bu, "%x#%llu",(unsigned int)__mess,num);\
 				string key(bu);\
 				__mess->setKey(key);\
 				DEBOUT("Incoming\n****************************************************\n",__mess->getIncBuffer() << "]\nkey [" << key)\
